lfs_manage/requirements-vendor.txt
django-compressor == 2.1.1
django-localflavor == 1.4.1
django-paypal == 0.3.6
django-portlets >= 1.5, < 1.6
django-postal == 0.96
django-reviews >= 1.2, < 1.3
lfs-contact >= 1.3, < 1.4
lfs-order-numbers >= 1.2, < 1.3
lfs-paypal >= 1.4, < 1.5
lfs-theme >= 0.11, < 0.12
django-storages == 1.6.5
google-cloud == 0.27.0
requests == 2.18.4